% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{titles_from}
\alias{titles_from}
\alias{links_from}
\alias{published_from}
\alias{sources_from}
\alias{shares_from}
\alias{entities_from}
\alias{highlights_from}
\alias{snippets_from}
\title{Restructure retrieved news}
\usage{
titles_from(..., format = c("tibble", "list"))

links_from(..., format = c("tibble", "list"))

published_from(..., format = c("tibble", "list"))

sources_from(..., format = c("tibble", "list"))

shares_from(..., format = c("tibble", "list"))

entities_from(..., format = c("tibble", "list"))

highlights_from(..., format = c("tibble", "list"))

snippets_from(..., format = c("tibble", "list"))
}
\arguments{
\item{...}{Objects of class "cf_api_list" retrieved using
\code{\link{get_news}}}

\item{format}{String specifying whether to return data as a "tibble" or
"list". Defaults to "tibble".}
}
\description{
Transform features of articles retrieved from CivicFeed News API into tidy
tibbles or lists
}
\details{
Each function extracts a different feature of the articles
  retrieved. The functions can be used in combination to create your desired
  dataset.
}
\section{Functions}{
\itemize{
\item \code{titles_from}: Extracts titles from retrieved articles

\item \code{links_from}: Extracts url, thumbnail and screenshot links from retrieved
articles

\item \code{published_from}: Extracts unix timestamps for dates on which retrieved
articles were published

\item \code{sources_from}: Extracts source IDs and names from retrieved articles

\item \code{shares_from}: Extracts timestamped social media shares, reactions and
comments from retrieved articles

\item \code{entities_from}: Extracts entity IDs, names, types etc. from retrieved articles

\item \code{highlights_from}: Extracts highlights from retrieved articles

\item \code{snippets_from}: Extracts snippets from retrieved articles
}}

\examples{
\dontrun{
news1 = get_news("congress")
news2 = get_news("white house")
# Good
entities_from(news1, news2)
## creates tibble of entities found in articles about congress and the white house

sources_from(news1, news2, format = "list")
## creates list of sources of articles about congress and the white house

# Bad
sources_from(news1, news2, "list")
## format argument must be labelled}
}
\seealso{
\code{\link{get_news}} to retrieve news from CivicFeed
}
