% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{lookup_id}
\alias{lookup_id}
\title{Lookup IDs of news sources}
\usage{
lookup_id(q = NULL, key = NULL)
}
\arguments{
\item{q}{String identifying a query to be searched through news source names.}

\item{key}{String specifying a valid CivicFeed API key. If \code{NULL},
\code{get_news} will look for a relevant API key stored as environmental
variable "CF_KEY".}
}
\value{
A named list of sources relevant to the query and their IDs. A maximum of 10
  relevant sources are returned.
}
\description{
Find out the IDs of news sources by querying a source name. Use the IDs returned in
\code{\link{get_news}} requests.
}
\examples{
\dontrun{
# Good
sources = lookup_id("new york times")
## NY times is the 7th element of "sources"

# Bad
lookup_id("times")
## too broad a query, source ID of NY times may not show up in 10 results returned}
}
\references{
\url{https://developers.civicfeed.com/documentation}
}
