% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentioned.R
\name{mentioned}
\alias{mentioned}
\title{Filter news mentioning specific entities}
\usage{
mentioned(..., entities, titles = TRUE, published = TRUE, sources =
  TRUE, snippets = FALSE, highlights = FALSE, shares = FALSE)
}
\arguments{
\item{...}{Objects of class "cf_api_list" retrieved using
\code{\link{get_news}}.}

\item{entities}{Numeric vector of entity IDs OR character vector of entity
names. Character vector is case-insensitive. Advisable to check \code{term}
entries in output returned by \code{\link{entities_from}} to avoid
misnaming.}

\item{titles}{Logical indicating whether to include titles of articles in
tibble. Defaults to \code{TRUE}.}

\item{published}{Logical indicating whether to include unix timestamps for
dates articles were published in tibble. Defaults to \code{TRUE}.}

\item{sources}{Logical indicating whether to include sources of articles in
tibble. Defaults to \code{TRUE}.}

\item{snippets}{Logical indicating whether to include snippets of articles in
tibble. Defaults to \code{FALSE}.}

\item{highlights}{Logical indicating whether  to include highlights of
articles in tibble. Defaults to \code{FALSE}. If \code{TRUE}, note:
articles may have more than one highlight. Each highlight will occupy
separate row in the tibble.}

\item{shares}{Logical indicating whether to include information about social
media shares of articles in tibble. Defaults to \code{FALSE}.}
}
\value{
A tibble displaying features of articles mentioning specified
  entitites.
}
\description{
From article results retrieved using \code{\link{get_news}}, extract a subset
that mention entities of interest and create a tibble of the features of that
subset.
}
\examples{
\dontrun{
news1 = get_news("congress")
news2 = get_news("white house")
# Good
mentioned(news1, news2, entities = c("hillary clinton", "president donald trump"), snippets = TRUE)
## creates a tibble of titles, dates published, sources and snippets of articles about congress
or white house mentioning h. clinton or trump

# Bad
mentioned(news1, news2, entities = c("hilary clinton", "donald trump"), snippets = TRUE)
## no matches due to spelling error and naming disparity}
}
