% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Check number of news results matching a request}
\usage{
results(q = NULL, sources = NULL, from = NULL, to = NULL, key =
  NULL)
}
\arguments{
\item{q}{String identifying a query to be searched through specified data
sources.}

\item{sources}{Numeric vector of IDs of data source to search through. Use
\code{lookup_id} to retrieve IDs.}

\item{from}{String identifying the start date (e.g. 2018-10-01) of content to
search. Date is in UTC.}

\item{to}{String identifying the end date (e.g. 2018-10-01) of content to
search. Date is in UTC.}

\item{key}{String specifying a valid CivicFeed API key. If \code{NULL},
\code{get_news} will look for a relevant API key stored as environmental
variable "CF_KEY".}
}
\value{
An integer specifying the number of results corresponding to the
  request.
}
\description{
Find out the number of results matching a CivicFeed News API request. Note: \code{results}
itself sends a request to the CivicFeed News API, so specify a valid API key
or have one stored as environment variable "CF_KEY".
}
\examples{
\dontrun{
results(q = "congress", from = "2018-10-01", to = "2018-11-01")
## finds out how many articles about congress were posted between 2018-10-01 and 2018-11-01
}
}
\references{
\url{https://developers.civicfeed.com/documentation}
}
